ver 2.0

+++ controllers/checkout.controller.php
@@ -13,18 +13,32 @@ Class CheckoutController extends Controller{
     public function index(){
 
         $cart = Session::get('cart');
+        $cart_g = Session::get('cart_g');
+        if( empty($cart) ){
+            $this->data['g'] = true;
+        }
 
-        if( is_null($cart) or empty($cart) ){
+        if( (is_null($cart) or empty($cart)) && (is_null($cart_g) or empty($cart_g)) ){
             Router::redirect('/cart/');
         } else {
             $this->data['cart_goods'] = $this->model->getCartGoods();
 
             $order = array();
-            foreach ($this->data['cart_goods'] as $cart_good){
-               $order[] = '(' . trim($cart_good['id']) . ')' . trim($cart_good['gasket_kit']) . '-' . trim($cart[$cart_good['id']]);
+            if($this->data['g']){
+                $this->data['code'] = 3;
+                foreach ($this->data['cart_goods'] as $cart_good){
+                    $order[] =  trim($cart_good['id']) . '-' . trim($cart_g[$cart_good['id']]);
+                }
+                $order = implode('; ',$order);
+            } else {
+                $this->data['code'] = 0;
+                foreach ($this->data['cart_goods'] as $cart_good){
+                    $order[] = '(' . trim($cart_good['id']) . ')' . trim($cart_good['gasket_kit']) . '-' . trim($cart[$cart_good['id']]);
+                }
+                $order = implode('; ',$order);
             }
-            $order = implode('; ',$order);
-            //$order = $order . ' //' . Session::get('sum') . ' grn';
+
+
 
             $this->data['order'] = $order;
             $this->data['sum'] = Session::get('sum');
@@ -32,22 +46,31 @@ Class CheckoutController extends Controller{
 
         }
 
-        if( $_POST ){
+        if( $_POST['name'] && $_POST['phone'] && $_POST['body'] && $_POST['email']){
             if($this->model->saveOrder($_POST)){
 
                 $cart = Session::get('cart');
-                $this->model->resetQuantity($cart);
+                $cart_g = Session::get('cart_g');
 
-                Session::delete('cart');
+                if( !empty($cart)){
+                    $this->model->resetQuantity($cart);
+                    Session::delete('cart');
+                }
+
+                if( !empty($cart_g)){
+                    $this->model->resetGasketQuantity($cart_g);
+                    Session::delete('cart_g');
+                }
+
                 Session::delete('sum');
                 Session::set('success',true);
                 Router::redirect('/checkout/success#st');
 
 
+            } else {
+                Router::redirect('/static.html');
             }
         }
-
-
     }
 
     public function success(){
@@ -61,6 +84,13 @@ Class CheckoutController extends Controller{
     }
 
 
+    public function admin_gasket(){
+        $this->data['g_orders'] = array_reverse($this->model->getGasketOrdersList());
+        $this->data['sum'] = 0;
+        foreach($this->data['g_orders'] as $order){
+            $this->data['sum'] = $this->data['sum'] + $order['sum'];
+        }
+    }
 
     public function admin_index(){
